# specs/001-backup-system/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: BackupChrono API
  description: Central pull backup system with device-centric organization
  version: 1.0.0
  contact:
    name: BackupChrono
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: https://backup.example.com/api
    description: Production server
...
paths:
...
  /devices:
    get:
      tags: [devices]
      summary: List all devices
      operationId: listDevices
      responses:
        '200':
          description: List of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
    
    post:
      tags: [devices]
      summary: Create a new device
      operationId: createDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreate'
      responses:
        '201':
          description: Device created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          $ref: '#/components/responses/BadRequest'
  /devices/{deviceId}:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    
    get:
      tags: [devices]
      summary: Get device by ID
      operationId: getDevice
      responses:
        '200':
          description: Device details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      tags: [devices]
      summary: Update device
      operationId: updateDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdate'
      responses:
        '200':
          description: Device updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      tags: [devices]
      summary: Delete device
      operationId: deleteDevice
      responses:
        '204':
          description: Device deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /devices/{deviceId}/test-connection:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    
    post:
      tags: [devices]
      summary: Test device connection
      operationId: testDeviceConnection
      responses:
        '200':
          description: Connection test result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  latency:
                    type: integer
                    description: Connection latency in milliseconds
        '404':
          $ref: '#/components/responses/NotFound'
  /devices/{deviceId}/wake:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    
    post:
      tags: [devices]
      summary: Send Wake-on-LAN packet to device
      operationId: wakeDevice
      responses:
        '200':
          description: WOL packet sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: WOL not enabled for this device
        '404':
          $ref: '#/components/responses/NotFound'
  /devices/{deviceId}/backups:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    
    get:
      tags: [devices, backups]
      summary: Get backup history for device
      operationId: getDeviceBackups
      responses:
        '200':
          description: Backup history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Backup'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== SHARES ====================
  /devices/{deviceId}/shares:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    
    get:
      tags: [shares]
      summary: List shares for a device
      operationId: listShares
      responses:
        '200':
          description: List of shares
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Share'
        '404':
          $ref: '#/components/responses/NotFound'
    
    post:
      tags: [shares]
      summary: Create a new share
      operationId: createShare
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShareCreate'
      responses:
        '201':
          description: Share created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Share'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /backups:
    get:
      tags: [backups]
      summary: List all backups
      operationId: listBackups
      parameters:
        - name: deviceId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by device ID
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
          description: Maximum number of results
      responses:
        '200':
          description: List of backups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Backup'
...
components:
  parameters:
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Device:
      type: object
      required: [id, name, protocol, host, username]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 100
        protocol:
          type: string
          enum: [SMB, SSH, Rsync]
        host:
          type: string
          minLength: 1
          maxLength: 255
        port:
          type: integer
          minimum: 1
          maximum: 65535
        username:
          type: string
          minLength: 1
          maxLength: 100
        wakeOnLanEnabled:
          type: boolean
          default: false
        wakeOnLanMacAddress:
          type: string
          pattern: '^([0-9A-Fa-f]{2}[:-]){5}([0-9A-Fa-f]{2})$'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DeviceDetail:
      allOf:
        - $ref: '#/components/schemas/Device'
        - type: object
          properties:
            schedule:
              $ref: '#/components/schemas/Schedule'
            retentionPolicy:
              $ref: '#/components/schemas/RetentionPolicy'
            includeExcludeRules:
              $ref: '#/components/schemas/IncludeExcludeRules'
            shares:
              type: array
              items:
                $ref: '#/components/schemas/Share'
            lastBackup:
              $ref: '#/components/schemas/Backup'
    DeviceCreate:
      type: object
      required: [name, protocol, host, username, password]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        protocol:
          type: string
          enum: [SMB, SSH, Rsync]
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        password:
          type: string
          writeOnly: true
          description: Device password (write-only, never returned in responses)
        wakeOnLanEnabled:
          type: boolean
        wakeOnLanMacAddress:
          type: string
        schedule:
          $ref: '#/components/schemas/Schedule'
        retentionPolicy:
          $ref: '#/components/schemas/RetentionPolicy'
        includeExcludeRules:
          $ref: '#/components/schemas/IncludeExcludeRules'
    DeviceUpdate:
      type: object
      properties:
        name:
          type: string
        protocol:
          type: string
          enum: [SMB, SSH, Rsync]
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        password:
          type: string
          writeOnly: true
          description: Device password (write-only, never returned in responses)
        wakeOnLanEnabled:
          type: boolean
        wakeOnLanMacAddress:
          type: string
        schedule:
          $ref: '#/components/schemas/Schedule'
        retentionPolicy:
          $ref: '#/components/schemas/RetentionPolicy'
        includeExcludeRules:
          $ref: '#/components/schemas/IncludeExcludeRules'

    # ========== SHARE SCHEMAS ==========
    Share:
      type: object
      required: [id, deviceId, name, path, enabled]
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 100
        path:
          type: string
          minLength: 1
          maxLength: 500
        enabled:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ShareCreate:
      type: object
      required: [name, path]
      properties:
        name:
          type: string
        path:
          type: string
        enabled:
          type: boolean
        schedule:
          $ref: '#/components/schemas/Schedule'
        retentionPolicy:
          $ref: '#/components/schemas/RetentionPolicy'
        includeExcludeRules:
          $ref: '#/components/schemas/IncludeExcludeRules'
...
    Backup:
      type: object
      properties:
        id:
          type: string
          description: Unique backup identifier
        deviceId:
          type: string
          format: uuid
        shareId:
          type: string
          format: uuid
          description: Share ID for share-level backups (null for device-level)
        deviceName:
          type: string
        shareName:
          type: string
          description: Share name for share-level backups (null for device-level)
        timestamp:
          type: string
          format: date-time
        status:
          type: string
          enum: [Success, Partial, Failed]
        sharesPaths:
          type: object
          description: Share names to paths mapping (all shares for device-level, single entry for share-level)
          additionalProperties:
            type: string
        fileStats:
          type: object
          description: File statistics
          properties:
            new:
              type: integer
              description: Number of new files
            changed:
              type: integer
              description: Number of changed files
            unmodified:
              type: integer
              description: Number of unmodified files
        dataStats:
          type: object
          description: Data processing statistics
          properties:
            added:
              type: integer
              format: int64
              description: Bytes of new data added
            processed:
              type: integer
              format: int64
              description: Total bytes processed
        duration:
          type: string
          description: ISO 8601 duration
        errorMessage:
          type: string
        createdByJobId:
          type: string
          format: uuid
          description: Backup job that created this backup
...
    Schedule:
      type: object
      required: [cronExpression]
      properties:
        cronExpression:
          type: string
          example: "0 2 * * *"
        timeWindowStart:
          type: string
          format: time
          example: "02:00"
        timeWindowEnd:
          type: string
          format: time
          example: "06:00"
    RetentionPolicy:
      type: object
      properties:
        keepLatest:
          type: integer
          minimum: 0
          default: 7
        keepDaily:
          type: integer
          minimum: 0
          default: 7
        keepWeekly:
          type: integer
          minimum: 0
          default: 4
        keepMonthly:
          type: integer
          minimum: 0
          default: 12
        keepYearly:
          type: integer
          minimum: 0
          default: 3
    IncludeExcludeRules:
      type: object
      properties:
        excludePatterns:
          type: array
          items:
            type: string
          example: ["*.tmp", "node_modules/"]
        excludeRegex:
          type: array
          items:
            type: string
        includeOnlyRegex:
          type: array
          items:
            type: string
        excludeIfPresent:
          type: array
          items:
            type: string
          example: [".nobackup"]

    # ========== CONFIGURATION SCHEMAS ==========
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true